+++ src/utils/vmem.c
@@ -71,7 +71,11 @@ static inline void* mmap_allocate(Vmem *vmem, size_t to_allocate)
 #endif
 	void *ptr = ((uint8_t *)vmem->ptr) + vmem->allocated;
 	vmem->allocated = allocated_after;
-	ASSERT0(vmem->size > vmem->allocated);
+	if (vmem->size < allocated_after)
+	{
+		error_exit("Error: The compiler ran out of memory! Over %u MB was allocated from a single memory arena, perhaps "
+				   "you called some recursive macro?", (unsigned)(vmem->size / (1024 * 1204)));
+	}
 	return ptr;
 }